requirements.txt
aiohttp==3.8.3
aiosignal==1.3.1
asgiref==3.6.0
async-timeout==4.0.2
attrs==22.2.0
black==23.1.0
charset-normalizer==2.1.1
click==8.1.3
Django==4.1.6
django-cors-headers==3.13.0
djangorestframework==3.14.0
frozenlist==1.3.3
idna==3.4
isort==5.12.0
multidict==6.0.4
mypy-extensions==1.0.0
packaging==23.0
pathspec==0.11.0
platformdirs==3.0.0
pytz==2022.7.1
sqlparse==0.4.3
tomli==2.0.1
yarl==1.8.2
